plugin_next_batch should return success when batch < max batch size
When using the NextBatch wrapper, NextBatch calls the provided next
function until a non-SSPluginSuccess value is returned.

There was a bug in this, where a timeout from the underlying next
function would be passed up to the framework, even if there were some
pending events (think fewer than max batch size).

The fix is to change a timeout to success if there are any events to
return.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>