new: cgo.Handle polyfill
Remove `state` pkg in favor a custom `cgo.Handle` implementation
optmized for performance in use cases related to plugins.
Using an atomic index with a fixed-size array is still thread-safe.
The hard limit of 32 handles is acceptable, since the intended usage
is only to pass opaque pointers holding the plugin states.
We can consider increasing that MaxHandle constant in the future.

Co-authored-by: Jason Dellaluce <jasondellaluce@gmail.com>
Co-authored-by: Federico Di Pierro <nierro92@gmail.com>
Signed-off-by: Leonardo Grasso <me@leonardograsso.com>