refactor: add type-agnostic SetValue in ExtractRequest
This removes the need of a SetXXXValue method for each supported field type. This leads to a cleaner codebase during plugin developers, and handles type casting internally.
The type-agnostic SetValue method does not make use of reflection, but instead uses the field type info provided by the framework to implement the internal type casting, which should be acceptable from a performance standpoint.
If type casting fails a panic is triggered, communicating the point of failure to plugin developers more explicitly.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>